+++ distributions/univariate/continuous/ExponentialRand.h
@@ -17,8 +17,8 @@
 class RANDLIBSHARED_EXPORT ExponentialRand : public GammaRand
 {
     /// Tables for ziggurat
-    static double stairWidth[257], stairHeight[256];
-    static double constexpr x1 = 7.69711747013104972;
+    static long double stairWidth[257], stairHeight[256];
+    static long double constexpr x1 = 7.69711747013104972l;
     static bool dummy;
     static bool SetupTables();
